feat: UI design and created upsert user functionality

+++ app/src/main/java/com/ham/activitymonitorapp/data/entities/Exercise.kt
@@ -9,8 +9,8 @@ import java.sql.Timestamp
 
 @Entity(tableName = "exercises")
 data class Exercise(
-    @PrimaryKey(autoGenerate = true) val exerciseId: Int = 0,
-    @ColumnInfo(name = "userId") val userId: Int,
+    @PrimaryKey(autoGenerate = true) val exerciseId: Long = 0,
+    @ColumnInfo(name = "userId") val userId: Long,
     @TypeConverters(DateTypeConverter::class)
     @ColumnInfo(name = "startTime") val startTime: Timestamp,
     @ColumnInfo(name = "duration") var duration: Int,
