+++ app/src/main/java/com/ham/activitymonitorapp/exceptions/HrNotFoundException.kt
@@ -1,3 +1,3 @@
 package com.ham.activitymonitorapp.exceptions
 
-class HrNotFoundException(hrId: Int) : Exception("Hr with hrId $hrId not found")
+class HrNotFoundException(hrId: Long) : Exception("Hr with hrId $hrId not found")
